Changed the Linux BACnet/IP init to use NetMask.
The function was modified to calculate the broadcast address from IP and netmask instead of using SIOCGIFBRDADDR. In some cases it is possible that the ioctl is successful, but the returned address is 0 (e.g. search for Bcast: 0.0.0.0). For some reason in Linux the local loopback device answers from 0.0.0.0 address. So messages broadcast to that address are received from 127.0.0.1 which can possibly create a broadcast loop. This has nothing to do with NAT, but makes the stack more robust. Thank you, Sami Pietikäinen, for the contribution!

git-svn-id: svn://svn.code.sf.net/p/bacnet/code/trunk@2808 1e814d3a-8a88-4260-be86-c0dc25669c1e